Move git commands implementations under Git/Commands
