+++ GitCommands/Git/Commands/GitCheckoutBranchCmd.cs
@@ -2,7 +2,7 @@
 using GitExtUtils;
 using GitUIPluginInterfaces;
 
-namespace GitCommands.Git
+namespace GitCommands.Git.Commands
 {
     public enum CheckoutNewBranchMode
     {
