+++ GitUI/CommandsDialogs/FormDeleteRemoteBranch.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 using GitCommands;
-using GitCommands.Git;
+using GitCommands.Git.Commands;
 using GitUI.HelperDialogs;
 using GitUI.Script;
 using GitUIPluginInterfaces;
