+++ UnitTests/GitCommands.Tests/Git/Tag/GitTagControllerTest.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.IO.Abstractions;
 using System.Windows.Forms;
+using GitCommands.Git.Commands;
 using GitCommands.Git.Tag;
 using GitUIPluginInterfaces;
 using NSubstitute;
